convert: Fake array size for zero-sized arrays, e.g. char name[0].
Developers often want to have structs with
elements whose size is determined at allocation time.
They therefore use patterns like this:
struct foo {
  ...
  char bar[0]
};
Since a ptr of struct foo usually points to the beginning
of the memory area, the remaining memory can be accessed
via bar.
This patterns makes it impossible to map memory accesses to
that area to an element of struct foo.
We therefore fake the size of array bar. For us, it will be 1MiByte.
1MiByte should be enough for everyone. :)

+++ queries/db-scheme.sql
@@ -64,7 +64,7 @@ CREATE TABLE `structs_layout` (
   `data_type_name` varchar(255) NOT NULL,			-- Describes the type of a member
   `member_name_id` int(11) UNSIGNED NOT NULL,		-- The id of the member
   `offset` smallint(11) UNSIGNED NOT NULL,	-- The offset in bytes from the beginning of a struct
-  `size` smallint(11) UNSIGNED NOT NULL,	-- The size in bytes of a member
+  `size` int(11) UNSIGNED NOT NULL,	-- The size in bytes of a member
   KEY (`data_type_id`, `offset`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8
 ;
